begin to refactoring and upgrading
